WARNINGS WAR Z: sign-compare warnings are now treated as errors
All warnings triggered by comparisons of unsigned with signed types have been
(once again) removed according to both GCC and Clang.

~150 had been added in the last year alone, and probably 75% came from 3
files.  PLEASE do not add warnings to Rosetta!  It is againt our coding con-
ventions and 99% of the time is bad coding.  If you do not understand what a
sign-compare warning is or if you would like advice on how to quench warnings
in general, see:

wiki.rosettacommons.org/index.php/List_of_Common_Warnings_and_How_to_Avoid_Them

As a rule of thumb: DO NOT USE int unless you actually want negative values.
If your for loop starts at 0 or 1 and counts up, use core::uint.  If your
variable stores a count or size use core::Size.

If you really need to compare a signed type with an unsigned one, static_cast
the unsigned type to the signed type, but 99% of the time it means you are
doing something wrong.

(A good number of these static_casts are forced on us because much of the old
code -- FoldTree, Pose, etc. -- is still using int where it should be using
core::Size.  I intend to refactor this at a later date, but for now, I'm shut
off the warnings by static_casting.)

This push also sets the sign-compare warning to be an error in GCC and Clang
builds.  If you compare a signed to an unsigned integer from this point on-
ward, Rosetta WILL NOT COMPILE on your machine (unless you are not using GCC
or Clang) and it will break the build on the test server.

(We decided to do this as a group at last MiniCON; we've only just got around
to making it happen.)

Note: It is possible that GCC on the test server will catch a sign-compare
that neither GCC nor Clang on my machine (Kubunutu 12) did.  I'll fix this
tomorrow morning if this occurs and will monitor the situation until the test
server agrees that all sign-compares are gone.

Note: If your code was laden with ints, it is possible that I broke some-
thing by accident that was not tested.  I tried to be careful to understand
what I was changing, but I could not always tell if an offset was meant to be
signed or not.  You may wish to double check any code that I have modified
with this push.  (And then go write a unit test.)

Thanks to Brian Weitzner for assistance and advice and to the rest of the
Fellowship of the WaRnINGs.

----------------------------------------
Unit test status: Pass
Integration test status: no changes

----
An additional 23 commit messages were squashed into this commit:
--
Warnings Removal: a few more sign-comapres caught by clang
--
Warnings Removal: turned on sign-compare as error for clang
--
Warnings Removal: more pilot app sign-compare warnings
--
Warnings Removal: more sign-compares from pilot apps
--
Warnings Removal: forgot about pilot apps; removed 2 more and more to follow
--
Warnings Removal: Werror=sign-compare shut off for unit tests, because the external libraries have tons of warnings
--
Warnings Removal: sign-compare warnings now treated as errors
--
Warnings Removal: removed last 2 sign-compare warnings caught on fourth build
--
Warnings Removal: removed another 12 sign-compare warnings caught on third build; 1 remains with gcc
--
Warnings Removal: fixed bugs and removed 4 more sign-compares caught on second build
--
Warnings Removal: removed 12 sign-compare warnings; 1 remains with gcc
--
Warnings Removal: removed 21 sign-compare warnings; 13 remain with gcc
--
Warnings Removal: removed 16 sign-compare warnings; 34 remain with gcc
--
Warnings Removal: removed 14 sign-compare warnings; 50 remain with gcc
--
Warnings Removal: removed 26 sign-compare warnings; 64 remain with gcc
--
Warnings Removal: removed 17 sign-compare warnings; 90 remain with gcc
--
mergin in master
--
merging in master
--
reverting dummy commit
--
dummy commit #2
--
It didn't work. reverting
--
dummy commit to test if git upgrade fixed -m issue
--
spelling error in template file
--

Old repository SHA1: 76243b4cc80bc4f0dd368723104b70013e80a1e6

+++ source/src/protocols/hybridization/ChunkTrialMover.cc
@@ -107,7 +107,7 @@ ChunkTrialMover::ChunkTrialMover(
 	highest_tmpl_resnum_ = 0;
 	for (core::Size i_template=1; i_template<=template_poses_.size(); ++i_template) {
 		for (core::Size ires=1; ires<=template_poses_[i_template]->total_residue(); ++ires) {
-			if (template_poses_[i_template]->pdb_info()->number(ires) > highest_tmpl_resnum_) {
+			if (template_poses_[i_template]->pdb_info()->number(ires) > static_cast<int>(highest_tmpl_resnum_)) {
 				highest_tmpl_resnum_ = template_poses_[i_template]->pdb_info()->number(ires);
 			}
 		}
